% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarizeHap.R
\name{binarizeHap}
\alias{binarizeHap}
\title{Binarizing Haplotype Membership Data}
\usage{
binarizeHap(hap_mat, P1_name, P2_name, P1_dosages, P2_dosages)
}
\arguments{
\item{hap_mat}{a matrix of phased group membership (haplotypes) with row
corresponding to marker-bins and columns to individuals as generated by
`phaseHap()`.}

\item{P1_name}{a character vector with the name of parent 1}

\item{P2_name}{a character vector with the name of parent 2}

\item{P1_dosages}{an integer with haplotypes presence for parent 1. This
would be 1 for parent 1 data and 0 for parent 2 data.}

\item{P2_dosages}{an integer with haplotypes presence for parent 2. This
would be 0 for parent 1 data and 1 for parent 2 data.}
}
\value{
a matrix of haplotypes presence /absence with as row haplotypes in
physical bins and as columns individuals.
}
\description{
`binarizeHap()` is binarizing group membership (haplotypes) data into marker
like presence / absence of haplotype group per bin.
}
\examples{
P1_phased_hap_bin <- binarizeHap(P1_phased_hapmat,
  P1_name = "P1",
  P2_name = "P2",
  P1_dosage = 1,
  P2_dosage = 0
)
}
