% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPop.R
\name{clusterPop}
\alias{clusterPop}
\title{Clustering Individuals Based on Distance Matrix}
\usage{
clusterPop(
  dist_list,
  method = "ward.D2",
  plot = F,
  bins_list = NULL,
  pdf_path = NULL
)
}
\arguments{
\item{dist_list}{a list of a list of dissimilarity structure of class "dist"
as generated by `getDist()`.}

\item{method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC).}

\item{plot}{a logical value indicating if clustered variants matrix should be
plotted. Be award that if set to TRUE this will substantially increase
running time.}

\item{bins_list}{the list of variant matrix generated by `makeBins()` and used
by `getDist()`. Only necessary if @param plot is set to TRUE.}

\item{pdf_path}{a character vector with the path to use to save the plots as
pdf.}
}
\value{
a list of object of class "hclust" which describes the tree of
individuals produced by the clustering process.
}
\description{
`clusterPop()` is a wrapper around stats::hclust() used to cluster
individuals based on a list of dissimilarity structure as produced by
`getDist()`.
It also offers the time consuming option to plot the list of variant matrix
clustered per variant and individuals and save the output as pdf.
}
\examples{
P1_clust_list <- clusterPop(dist_P1_bins, plot = T,
bins_list = P1_bins, pdf_path = "./mypop_chr01_P1_clustering.pdf")
}
