% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSD.R
\name{formatSD}
\alias{formatSD}
\title{Format Phased Haplotype Membership for SmoothDescent}
\usage{
formatSD(P1_hap_mat, P2_hap_mat, P1_name, P2_name)
}
\arguments{
\item{P1_hap_mat}{a matrix of phased group membership (haplotypes) with row
corresponding to physical bins and columns to individuals as generated by
`phaseHap()` for parent 1.}

\item{P2_hap_mat}{a matrix of phased group membership (haplotypes) with row
corresponding to physical bins and columns to individuals as generated by
`phaseHap()` for parent 2.}

\item{P1_name}{a character vector with the name of parent 1.}

\item{P2_name}{a character vector with the name of parent 2.}
}
\value{
a list of 3 objects usable as input by SmoothDescent:
\itemize{
  \item"geno": matrix with markers on the rows, individuals on the columns
  with marker names as Rownames. As expected by `smooth_decent()` "geno"
  parameter.
  \item"hom": matrix with markers on the rows, homologue names on the columns
   as expected by `smooth_decent()` "homologue" parameter.
  \item"map": data.frame with at a columns "marker" containing marker name
  and "position" containing the physical position of the marker. As expected
  by `smooth_decent()` "map" parameter.}
}
\description{
`formatSD()` aggregate P1 and P2 group membership (haplotypes) data generated
by OutcrossSeqDiploidR and format it for SmoothDescent.
}
\examples{
SD_input <- formatSD(P1_phased_hap$hap_mat,
P2_phased_hap$hap_mat,
P1_name = "P1",
P2_name = "P2")
}
