% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist.R
\name{getDist}
\alias{getDist}
\title{Get Distance Between Individuals}
\usage{
getDist(bins_list, method = "euclidean")
}
\arguments{
\item{bins_list}{a list of matrix as generated by `makeBins()`.}

\item{method}{the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".
 Any unambiguous substring can be given.}
}
\value{
a list of dissimilarity structure between individuals in local bins of
class "dist".
}
\description{
`getDist()` is a wrapper around stats::dist(). It is applied to a list of
matrix to compute distance between individuals based on local variants.
}
\examples{
dist_P1_bins <- getDist(P1_bins, method = "euclidean")
}
