% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHap.R
\name{getHap}
\alias{getHap}
\title{Impute Local Haplotypes From Clustering Data}
\usage{
getHap(clust_list, k = 2)
}
\arguments{
\item{clust_list}{a list of object of class "hclust" as produced by
`clusterPop()`describing the tree of individuals produced by local
clustering.}

\item{k}{an integer with the desired number of groups (haplotypes) to
extract. This would be 2 for each parent of a diploid population.}
}
\value{
a list of vector corresponding to physical bins with group membership
 corresponding to local haplotypes.
}
\description{
`getHap()` is a wrapper around stats::cutree used to extract a given number
of clusters ,corresponding to local haplotypes (i.e. bin-markers alleles),
from a list of object of class "hclust" produced by `clusterPop()`
}
\examples{
P1_hap_list <- getHap(P1_clust_list, k=2)
}
