% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBins.R
\name{makeBins}
\alias{makeBins}
\title{Subdivide Chromosomal Variant Matrix in Bins}
\usage{
makeBins(parental_matrix, bin_size, snp_pos_colname = "pos")
}
\arguments{
\item{parental_matrix}{a parent specific chromosomal matrix as generated by
`recodeParentMatList()`.}

\item{bin_size}{an integer with the bin size to use in the same unit as the
marker position column of the parental_matrix.}

\item{snp_pos_colname}{a character vector with the column name of marker
position.}
}
\value{
a list of matrix corresponding the subdivision in bins of given
physical length of the input chromosomal matrix.
}
\description{
`makeBins()` subdivide a given chromosomal matrix into a list of bins matrix
of given physical length.
}
\examples{
P1_bins <- makeBins(parent_mat_list[['P1']], bin_size = 100000)
}
