% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualPhasing.R
\name{manualPhasing}
\alias{manualPhasing}
\title{Phase Local Haplotypes manually}
\usage{
manualPhasing(hap_mat, switch_marker)
}
\arguments{
\item{hap_mat}{a matrix of phased group membership (haplotypes) with row
corresponding to physical bins and columns to individuals as generated by
`phaseHap()`.}

\item{switch_marker}{a character vector corresponding to the bin name onward
from which phasing need to be flipped. The character vector should be part of
`rownames(hap_mat)`.}
}
\value{
a matrix of haplotype with row corresponding to physical bins
and columns to individuals with manually corrected phasing.
}
\description{
`manualPhasing()` is used to manual flip the phasing of a bin and all
subsequent bins when, upon visual inspection, a segment was wrongly phased by
`phaseHap()`.
}
\examples{
P1_phased_hapmat <- manualPhasing(P1_phased_hapmat,"p1_43.9_44_Mb")
}
