% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseHap.R
\name{phaseHap}
\alias{phaseHap}
\title{Phase Local Haplotypes to Minimize Recombination Frequency}
\usage{
phaseHap(hap_list)
}
\arguments{
\item{hap_list}{a list of vector corresponding to marker-bins with group
membership corresponding to local haplotypes as generated by `getHap()` and
preferably filtered using the information generated by `qualCheck()`.}
}
\value{
a list of 2 elements:
\itemize{
  \item "hap_mat" a matrix of phased group membership (haplotypes) with row
corresponding to marker-bins and columns to individuals.
  \item "recoded" a vector of logical values the same length as "hap_list"
indicating if the given bin phasing was flipped.}
}
\description{
`phaseHap()` is calculating recombination frequencies between physically
adjacent bins and minimize them by flipping the haplotype phasing if
necessary.
After a first round of phasing with `phaseHap()`, it is advised to visualize
the output, remove bins with aberrant haplotype data and, manually correct
phasing using `manualPhasing()` if necessary.
}
\examples{
P1_phased_hap <- phaseHap(P1_hap_list[P1_flt_bins])
}
