% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualCheck.R
\name{qualCheck}
\alias{qualCheck}
\title{Quality Check Local Haplotypes Imputation}
\usage{
qualCheck(
  hap_list,
  min_rec = 0.3,
  max_rec = 0.7,
  mv_av = 11,
  max_dist = 0.1,
  save_plot = F,
  pdf_path = NULL
)
}
\arguments{
\item{hap_list}{a list of vector corresponding to physical bins with group
membership corresponding to local haplotypes as generated by `getHap()`.}

\item{min_rec}{a numeric corresponding to the minimal recombination frequency
with previous and next bins above which a bin is flagged.}

\item{max_rec}{a numeric corresponding to the maximal recombination frequency
with previous and next bins bellow which a bin is flagged. This value
should be higher than 0.5 to cover for flipped phasing.}

\item{mv_av}{an integer corresponding to the number of bins to use for
calculating the moving average of transmission ratio distortion against which
the local transmission ration distortion will be compared.}

\item{max_dist}{a numeric corresponding to the maximal absolute difference
between the moving average and the local transmission ratio distortion above
which a bin is flagged.}

\item{save_plot}{a logical value indicating if summary quality plots should
be saved.}

\item{pdf_path}{a character vector with the path to use to save the plots as
pdf.}
}
\value{
a dataframe with rows corresponding to bins (elements of
parameter "hap_list") and the following columns:
\itemize{
  \item"bin_i": bin of interest
  \item "bin_h": previous bin
  \item"bin_j": next bin
  \item"pos_bin_i": physical position of bin of interest
  \item"rec_hi": recombination frequency between bin h and i
  \item"rec_ij": recombination frequency between bin i and j
  \item"loc_dist": absolute value of bin i deviation from the expected 1:1
    transmission ratio distortion.
  \item"ma_dist": moving average of deviation from the expected 1:1
    transmission ratio distortion.
  \item"quality": character vector flagging aberrant bins. Either
    "drop (high rec)", "drop (high dist)" or "keep".}
}
\description{
`qualCheck()` calculates recombination frequencies between adjacent bins and
local distortions from the expected 1:1 transmission ratio.Those values
are then compared with user imputed thresholds for recombination and a
moving average for distortion to flag bins with data judged unlikely to be
of biological origin. The function also offers the option to save summary
plots as pdf.
}
\examples{
P1_qual_check <- qualCheck(P1_hap_list,
  min_rec = 0.30,
  max_rec = 0.70, mv_av = 11,
  max_dist = 0.15, save_plot = T,
  pdf_path = "./mypop_chr01_P1_imputation_quality_check.pdf"
)

P1_flt_bins <- P1_qual_check$bin_i[which(P1_qual_check$quality == "keep")]
}
