% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeParentMatList.R
\name{recodeParentMatList}
\alias{recodeParentMatList}
\title{Re-code Bi-parental Dosages Matrix}
\usage{
recodeParentMatList(dosages_matrix, P1_name, P2_name, snp_pos_colname = "pos")
}
\arguments{
\item{dosages_matrix}{an integer matrix with variant in rows
and individuals, parents and marker physical position (in bp) in columns.}

\item{P1_name}{a character vector with the sample names of parent 1.}

\item{P2_name}{a character vector with the sample names of parent 2.}

\item{snp_pos_colname}{a character vector with the column name of marker
position.}
}
\value{
return a list of 2 (male and female) integer matrix with with variant
 in rows and individuals, parents and marker position in columns.
}
\description{
`recodeParentMatList()` re-code a chromosomal bi-parental dosage matrix into
 2 parental integer matrix representing presence / absence data of parent
 specific variant encoded under their simplest form (e.g. 1x2 -> 1x0).
}
\examples{
parent_mat_list<- recodeParentMatList(dosages_matrix,
                                                "P1",
                                                "P2",
                                                "pos")
}
